user authentication and chat functionality
The User which includes Shelter and Adopter role were implemented on the client side using AXIOS to make request to the server URL API.
Firebase Realtime Database was used for the chat functionality.
When a user (adopter) logs in, the user is able to use the chat function to chat with an admin (shelter). The admin is notified of a user waiting in the chat room with a request to join.
React Bootstrap and Modal was used for login and register UX.